+++ python/nvstrings/nvtext.py
@@ -379,3 +379,38 @@ def scatter_count(strs, counts):
     if rtn is not None:
         rtn = nvs.nvstrings(rtn)
     return rtn
+
+
+def porter_stemmer_measure(strs, vowels="aeiou", y_char="y", devptr=0):
+    """
+    Calculates porter stemmer measure of each string
+
+    Parameters
+    ----------
+    strs : nvstrings
+        The strings for this operation
+
+    vowels: str
+     chars to consider as vowels
+     Default are 'aeiou'
+
+    y_char: str
+     characters to treat as 'y_char'
+     Default is 'y'
+
+    devptr : GPU memory pointer
+        Must be able to hold at least strs.size() of int32 values.
+
+    Examples
+    --------
+    >>> import nvstrings, nvtext
+    >>> s = nvstrings.to_device(["tr","trouble", "troubles"])
+    >>> n = nvtext.porter_stemmer_measure(s)
+    >>> print(n)
+    >>> [0, 1, 2]
+    """
+    if not isinstance(strs, nvs.nvstrings):
+        raise ValueError("strs must be nvstrings object")
+
+    rtn = pyniNVText.n_porter_stemmer_measure(strs, vowels, y_char, devptr)
+    return rtn